requirements.txt
# GitHub API
PyGithub>=2.1.1
requests>=2.31.0

# Jira API
jira>=3.5.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Date handling
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
responses>=0.23.0
